+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+v0.9.x 
+
+- new: micronaut provider
+- new: path transformation plugin
+
 v0.8.x - Renderer Plugins
 
 - new: gradle plugin